refactor(llm): move cmd/instinct/llm to internal/instinctllm — fix cross-cmd import
Blocker 2 from Phase 1 adversarial review: cmd/audit imported
github.com/petersimmons1972/engram/cmd/instinct/llm, a cross-cmd import
anti-pattern in Go.

Renamed package to instinctllm (not just llm) because internal/llm already
exists for a different concern (OpenAI-compatible chat client used by
summarize/consolidate). Distinct names prevent identifier collision if both
packages are ever imported together and make import diffs unambiguous.

- git mv preserves history
- All importers updated (cmd/audit, cmd/instinct, cmd/instinct/consolidator)
- Package doc updated to explain the rename and the naming distinction
- go test ./... -count=1 -race: all green
