+++ cmd/audit/main.go
@@ -9,7 +9,7 @@ import (
 	"os"
 	"time"
 
-	"github.com/petersimmons1972/engram/cmd/instinct/llm"
+	"github.com/petersimmons1972/engram/internal/instinctllm"
 )
 
 // Version is injected at build time via -ldflags "-X main.Version=<ver>".
@@ -40,7 +40,7 @@ func run() int {
 	// tolerates a missing backend), audit is a one-shot batch tool — failure
 	// should be loud and non-zero so cron jobs surface the problem.
 	os.Setenv("LLM_BACKEND", *llmBackend) //nolint:errcheck
-	client, err := llm.NewClient(llm.Config{Timeout: *timeout})
+	client, err := instinctllm.NewClient(instinctllm.Config{Timeout: *timeout})
 	if err != nil {
 		slog.Error("create LLM client", "backend", *llmBackend, "err", err)
 		return 1
@@ -55,7 +55,7 @@ func run() int {
 
 // runAudit fetches patterns, judges each, and writes the JSON report to out.
 // Extracted from run() so it can be tested without exec.
-func runAudit(base, token string, client llm.LLMClient, timeout time.Duration, out io.Writer) error {
+func runAudit(base, token string, client instinctllm.LLMClient, timeout time.Duration, out io.Writer) error {
 	if timeout == 0 {
 		timeout = defaultTimeout
 	}
