+++ internal/instinctllm/anthropic.go
@@ -1,4 +1,4 @@
-package llm
+package instinctllm
 
 import (
 	"bytes"
