+++ internal/instinctllm/factory.go
@@ -1,4 +1,4 @@
-package llm
+package instinctllm
 
 import (
 	"fmt"
