+++ internal/instinctllm/olla.go
@@ -1,4 +1,4 @@
-package llm
+package instinctllm
 
 import (
 	"bytes"
